Refactor calculator logic; put it inside Calculator component
